+++ client/web/src/search/home/SearchPageInput.tsx
@@ -3,7 +3,6 @@ import React, { useState, useCallback, useEffect, useMemo } from 'react'
 import { Form } from 'reactstrap'
 
 import { ActivationProps } from '@sourcegraph/shared/src/components/activation/Activation'
-import { Link } from '@sourcegraph/shared/src/components/Link'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { VersionContextProps } from '@sourcegraph/shared/src/search/util'
 import { SettingsCascadeProps, isSettingsValid } from '@sourcegraph/shared/src/settings/settings'
@@ -51,8 +50,6 @@ interface Props
     availableVersionContexts: VersionContext[] | undefined
     /** Whether globbing is enabled for filters. */
     globbing: boolean
-    /** Show the query builder link. */
-    showQueryBuilder: boolean
     /** A query fragment to appear at the beginning of the input. */
     queryPrefix?: string
     /** A query fragment to be prepended to queries. This will not appear in the input until a search is submitted. */
@@ -118,13 +115,6 @@ export const SearchPageInput: React.FunctionComponent<Props> = (props: Props) =>
                         autoFocus={showOnboardingTour ? shouldFocusQueryInput : props.autoFocus !== false}
                     />
                 </div>
-                {props.showQueryBuilder && (
-                    <div className="search-page__input-sub-container">
-                        <Link className="btn btn-link btn-sm pl-0" to="/search/query-builder">
-                            Query builder
-                        </Link>
-                    </div>
-                )}
                 <QuickLinks quickLinks={quickLinks} className="search-page__input-sub-container" />
                 <Notices className="my-3" location="home" settingsCascade={props.settingsCascade} />
             </Form>